feat: Add wallet birthday to `CreateParams`
There has been discussion offline on how to improve the experience for
recovering wallets using block-by-block scanning methods. A birthday has
come up multiple times, so I introduce one here in `CreateParams`.

I believe the `expect` may only be hit when attempting to over-write the
genesis block. If we want to return the error, it would require a break
in the error type. IMO this seems unnecessary, as a birthday should not
attempt to replace genesis.

This may be used immediately in `Wallet::latest_checkpoint` or similar.